# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Secure Authentication System API
  description: |
    A comprehensive authentication system with MFA, RBAC, and security features.
    
    ## Features
    - User registration and authentication
    - Multi-factor authentication (MFA)
    - Role-based access control (RBAC)
    - Session management
    - Activity logging
    - Challenge/CAPTCHA system
    
    ## Authentication
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-token>
    ```
  version: 1.0.0
  contact:
    name: API Support
    email: support@yourcompany.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.yourcompany.com
    description: Production server
...
paths:
...
  /api/auth/login:
    post:
      summary: User login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - passcode
              properties:
                username:
                  type: string
                  example: john_doe
                passcode:
                  type: string
                  example: SecurePass123!
                loginMethod:
                  type: string
                  enum: [direct, telegram]
                  default: direct
                challengeId:
                  type: string
                  format: uuid
                  description: Required if challenge was requested
                challengeAnswer:
                  type: string
                  description: Answer to the challenge
      responses:
        '200':
          description: Login successful or MFA required
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                  requireMFA:
                    type: boolean
                  userId:
                    type: string
                    format: uuid
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many failed attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        registrationMethod:
          type: string
          enum: [telegram, whatsapp]
        isVerified:
          type: boolean
        mfaEnabled:
          type: boolean
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: object
              nullable: true
            timestamp:
              type: string
              format: date-time
...